+++ api/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@ibm/itest",
-  "displayName": "IBM i Testing CLI",
-  "description": "IBM i Testing CLI",
+  "name": "@ibm/ibmi-testing-api",
+  "displayName": "IBM i Testing API",
+  "description": "IBM i Testing API",
   "author": {
     "name": "IBM",
     "url": "https://github.com/IBM"
@@ -27,12 +27,7 @@
     "Testing"
   ],
   "main": "./dist/index.js",
-  "types": "./dist/src/index.d.ts",
-  "bin": {
-    "itest": "./dist/index.js"
-  },
   "scripts": {
-    "test": "vitest",
     "webpack:dev": "webpack --mode none --config ./webpack.config.js",
     "webpack": "webpack --mode production --config ./webpack.config.js",
     "local": "npm run webpack:dev && npm i -g",
@@ -40,22 +35,21 @@
   },
   "dependencies": {
     "ansi-colors": "^4.1.3",
-    "commander": "^13.1.0",
-    "inquirer": "^12.7.0",
-    "node-ssh": "^13.1.0",
-    "ora": "^8.2.0",
+    "lodash": "^4.17.21",
     "table": "^6.9.0",
+    "tmp": "^0.2.3",
+    "unzipper": "^0.11.6",
+    "xml2js": "^0.6.2",
     "vscode-ibmi": "github:halcyon-tech/vscode-ibmi",
-    "vscode-rpgle": "github:halcyon-tech/vscode-rpgle",
-    "xml2js": "^0.6.2"
+    "vscode-rpgle": "github:halcyon-tech/vscode-rpgle"
   },
   "devDependencies": {
-    "@types/ssh2": "^1.15.5",
+    "@types/lodash": "^4.17.16",
+    "@types/tmp": "^0.2.6",
+    "@types/unzipper": "^0.10.11",
     "@types/xml2js": "^0.4.14",
-    "copy-webpack-plugin": "^13.0.0",
     "ts-loader": "^9.4.4",
     "typescript": "^5.8.2",
-    "vitest": "^3.0.8",
     "webpack": "^5.24.3",
     "webpack-cli": "^4.5.0"
   }
