--- ibmi-testing.code-workspace
@@ -1,13 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "."
-		},
-		{
-			"path": "src/cli"
-		}
-	],
-	"settings": {
-		"typescript.tsc.autoDetect": "off"
-	}
-}
