+++ package.json
@@ -237,20 +237,16 @@
     "vscode-rpgle": "github:halcyon-tech/vscode-rpgle",
     "ansi-colors": "^4.1.3",
     "compare-versions": "^6.1.1",
-    "lodash": "^4.17.21",
     "node-fetch": "^3.3.2",
     "octokit": "^4.1.2",
     "tmp": "^0.2.3",
-    "unzipper": "^0.11.6",
     "xml2js": "^0.6.2"
   },
   "devDependencies": {
     "@halcyontech/vscode-ibmi-types": "^2.16.0",
-    "@types/lodash": "^4.17.16",
     "@types/mocha": "^10.0.7",
     "@types/node": "20.x",
     "@types/tmp": "^0.2.6",
-    "@types/unzipper": "^0.10.11",
     "@types/vscode": "^1.96.0",
     "@types/xml2js": "^0.4.14",
     "@typescript-eslint/eslint-plugin": "^7.14.1",
